refactor get graphql link
