+++ src/ActionModal/utils/ActionManager.js
@@ -3,7 +3,7 @@ import { getSigner } from "./signer"
 import transaction from "./transactionTypes"
 import { uatoms } from "scripts/num"
 import { toMicroDenom } from "src/scripts/common"
-import { getURLParams } from "scripts/url"
+import { getGraphqlHost } from "scripts/url"
 import {
   getMessage,
   getMultiMessage,
@@ -77,16 +77,11 @@ export default class ActionManager {
 
     const command = payload.simulate ? "estimate" : "broadcast"
 
-    // TODO refactor and put into config.js
-    const graphqlHost = urlParams =>
-      (urlParams.graphql ? decodeURIComponent(urlParams.graphql) : false) ||
-      config.graphqlHost
-    const urlParams = getURLParams(window)
+    const graphqlHost = getGraphqlHost()
 
-    return fetch(
-      `${graphqlHost(urlParams)}/transaction/${command}`,
-      options
-    ).then(r => r.json())
+    return fetch(`${graphqlHost}/transaction/${command}`, options).then(r =>
+      r.json()
+    )
   }
 
   async simulateTxAPI(context, type, txProps, memo) {
