+++ src/gql/apollo.js
@@ -8,55 +8,12 @@ import { InMemoryCache } from "apollo-cache-inmemory"
 import { split } from "apollo-link"
 import { getMainDefinition } from "apollo-utilities"
 import VueApollo from "vue-apollo"
-import config from "src/../config"
+import { getGraphqlHost } from "scripts/url"
 
 Vue.use(VueApollo)
 
-// persistent api
-// setting api url in localStorage
-const checkPersistentAPI = urlParams => {
-  if (
-    typeof urlParams.iwouldliketochangetheapipersistentlyandiknowwhatido !==
-    "undefined"
-  ) {
-    // removing presistent api on false value
-    if (
-      !urlParams.iwouldliketochangetheapipersistentlyandiknowwhatido ||
-      urlParams.iwouldliketochangetheapipersistentlyandiknowwhatido == "false"
-    ) {
-      localStorage.removeItem(`persistentapi`)
-      return false
-    }
-    // setting presistent api
-    if (urlParams.iwouldliketochangetheapipersistentlyandiknowwhatido) {
-      localStorage.setItem(
-        `persistentapi`,
-        decodeURIComponent(
-          urlParams.iwouldliketochangetheapipersistentlyandiknowwhatido
-        )
-      )
-      return decodeURIComponent(
-        urlParams.iwouldliketochangetheapipersistentlyandiknowwhatido
-      )
-    }
-  }
-
-  if (localStorage.getItem(`persistentapi`)) {
-    console.warn(
-      "Your API version differ from normal. Query ?iwouldliketochangetheapipersistentlyandiknowwhatido=false to reset"
-    )
-    return localStorage.getItem(`persistentapi`)
-  }
-  return false
-}
-
-const graphqlHost = urlParams =>
-  checkPersistentAPI(urlParams) ||
-  (urlParams.api ? decodeURIComponent(urlParams.api) : false) ||
-  config.graphqlHost
-
-const makeHttpLink = urlParams => {
-  const host = graphqlHost(urlParams)
+const makeHttpLink = () => {
+  const host = getGraphqlHost()
   const uri = host
 
   // We create a createPersistedQueryLink to lower network usage.
@@ -70,14 +27,14 @@ const makeHttpLink = urlParams => {
   )
 }
 
-const makeWebSocketLink = urlParams => {
-  const host = graphqlHost(urlParams)
+const makeWebSocketLink = () => {
+  const host = getGraphqlHost()
   const uri = `${host.replace("http", "ws")}/graphql`
   console.log("ws", uri)
   return new WebSocketLink({ uri })
 }
 
-const createApolloClient = urlParams => {
+const createApolloClient = () => {
   const link = split(
     ({ query }) => {
       const definition = getMainDefinition(query)
@@ -86,8 +43,8 @@ const createApolloClient = urlParams => {
         definition.operation === "subscription"
       )
     },
-    makeWebSocketLink(urlParams),
-    makeHttpLink(urlParams)
+    makeWebSocketLink(),
+    makeHttpLink()
   )
 
   const cache = new InMemoryCache()
@@ -100,9 +57,9 @@ const createApolloClient = urlParams => {
   })
 }
 
-export const createApolloProvider = urlParams => {
+export const createApolloProvider = () => {
   return new VueApollo({
-    defaultClient: createApolloClient(urlParams),
+    defaultClient: createApolloClient(),
     defaultOptions: {
       // apollo options applied to all queries in components
       $query: {
